refactor: centralize TTS audio handling
