+++ client/src/components/Audio/TTS.tsx
@@ -2,7 +2,10 @@
 import { useEffect, useMemo } from 'react';
 import { useRecoilValue } from 'recoil';
 import type { TMessageAudio } from '~/common';
-import { useLocalize, useTTSBrowser, useTTSExternal } from '~/hooks';
+import { useLocalize, useTTSBrowser } from '~/hooks';
+import { useSetRecoilState } from 'recoil';
+import { parseTextParts } from 'librechat-data-provider';
+import { ttsRequestAtom } from '~/store/audio';
 import { VolumeIcon, VolumeMuteIcon, Spinner } from '~/components';
 import { logger } from '~/utils';
 import store from '~/store';
@@ -86,77 +89,22 @@ export function BrowserTTS({ isLast, index, messageId, content, className }: TMe
 
 export function ExternalTTS({ isLast, index, messageId, content, className }: TMessageAudio) {
   const localize = useLocalize();
-  const playbackRate = useRecoilValue(store.playbackRate);
+  const setRequest = useSetRecoilState(ttsRequestAtom);
 
-  const { toggleSpeech, isSpeaking, isLoading, audioRef } = useTTSExternal({
-    isLast,
-    index,
-    messageId,
-    content,
-  });
-
-  const renderIcon = (size: string) => {
-    if (isLoading === true) {
-      return <Spinner size={size} />;
-    }
-
-    if (isSpeaking === true) {
-      return <VolumeMuteIcon size={size} />;
-    }
-
-    return <VolumeIcon size={size} />;
+  const handleClick = () => {
+    const messageContent = content ?? '';
+    const text = typeof messageContent === 'string' ? messageContent : parseTextParts(messageContent);
+    setRequest({ text });
   };
 
-  useEffect(() => {
-    const messageAudio = document.getElementById(`audio-${messageId}`) as HTMLAudioElement | null;
-    if (!messageAudio) {
-      return;
-    }
-    if (playbackRate != null && playbackRate > 0 && messageAudio.playbackRate !== playbackRate) {
-      messageAudio.playbackRate = playbackRate;
-    }
-  }, [audioRef, isSpeaking, playbackRate, messageId]);
-
-  logger.log(
-    'MessageAudio: audioRef.current?.src, audioRef.current',
-    audioRef.current?.src,
-    audioRef.current,
-  );
-
   return (
-    <>
-      <button
-        className={className}
-        onClickCapture={() => {
-          if (audioRef.current) {
-            audioRef.current.muted = false;
-          }
-          toggleSpeech();
-        }}
-        type="button"
-        title={isSpeaking === true ? localize('com_ui_stop') : localize('com_ui_read_aloud')}
-      >
-        {renderIcon('19')}
-      </button>
-      <audio
-        ref={audioRef}
-        controls
-        preload="none"
-        controlsList="nodownload nofullscreen noremoteplayback"
-        style={{
-          position: 'absolute',
-          overflow: 'hidden',
-          display: 'none',
-          height: '0px',
-          width: '0px',
-        }}
-        src={audioRef.current?.src}
-        onError={(error) => {
-          logger.error('Error fetching audio:', error);
-        }}
-        id={`audio-${messageId}`}
-        autoPlay
-      />
-    </>
+    <button
+      className={className}
+      onClick={handleClick}
+      type="button"
+      title={localize('com_ui_read_aloud')}
+    >
+      <VolumeIcon size="19" />
+    </button>
   );
 }
