+++ client/src/components/Audio/Voices.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { useRecoilState } from 'recoil';
 import type { Option } from '~/common';
-import { useLocalize, useTTSBrowser, useTTSExternal } from '~/hooks';
+import { useLocalize, useTTSBrowser } from '~/hooks';
+import { useVoicesQuery } from '~/data-provider';
 import { Dropdown } from '~/components/ui';
 import { logger } from '~/utils';
 import store from '~/store';
@@ -37,7 +38,7 @@ export function BrowserVoiceDropdown() {
 
 export function ExternalVoiceDropdown() {
   const localize = useLocalize();
-  const { voices = [] } = useTTSExternal();
+  const { data: voicesData = [] } = useVoicesQuery();
   const [voice, setVoice] = useRecoilState(store.voice);
 
   const handleVoiceChange = (newValue?: string | Option) => {
@@ -52,9 +53,9 @@ export function ExternalVoiceDropdown() {
     <div className="flex items-center justify-between">
       <div>{localize('com_nav_voice_select')}</div>
       <Dropdown
-        key={`external-voice-dropdown-${voices.length}`}
+        key={`external-voice-dropdown-${voicesData.length}`}
         value={voice ?? ''}
-        options={voices}
+        options={voicesData}
         onChange={handleVoiceChange}
         sizeClasses="min-w-[200px] !max-w-[400px] [--anchor-max-width:400px]"
         testId="ExternalVoiceDropdown"
