+++ client/src/hooks/Input/useTextToSpeech.ts
@@ -3,7 +3,6 @@ import { useRef, useMemo, useEffect, useState } from 'react';
 import { parseTextParts } from 'librechat-data-provider';
 import type { TMessageContentParts } from 'librechat-data-provider';
 import type { Option } from '~/common';
-import useTextToSpeechExternal from '~/hooks/Input/useTextToSpeechExternal';
 import useTextToSpeechBrowser from '~/hooks/Input/useTextToSpeechBrowser';
 import useGetAudioSettings from '~/hooks/Input/useGetAudioSettings';
 import useAudioRef from '~/hooks/Audio/useAudioRef';
@@ -39,52 +38,15 @@ const useTextToSpeech = (props?: TUseTextToSpeech) => {
     voices: voicesLocal,
   } = useTextToSpeechBrowser({ setIsSpeaking });
 
-  const {
-    generateSpeechExternal,
-    cancelSpeech: cancelSpeechExternal,
-    isLoading: isLoadingExternal,
-    voices: voicesExternal,
-  } = useTextToSpeechExternal({
-    setIsSpeaking,
-    audioRef,
-    messageId,
-    isLast,
-    index,
-  });
+  const generateSpeech = useMemo(() => generateSpeechLocal, [generateSpeechLocal]);
 
-  const generateSpeech = useMemo(() => {
-    const map = {
-      browser: generateSpeechLocal,
-      external: generateSpeechExternal,
-    };
+  const cancelSpeech = useMemo(() => cancelSpeechLocal, [cancelSpeechLocal]);
 
-    return map[textToSpeechEndpoint];
-  }, [generateSpeechExternal, generateSpeechLocal, textToSpeechEndpoint]);
-
-  const cancelSpeech = useMemo(() => {
-    const map = {
-      browser: cancelSpeechLocal,
-      external: cancelSpeechExternal,
-    };
-    return map[textToSpeechEndpoint];
-  }, [cancelSpeechExternal, cancelSpeechLocal, textToSpeechEndpoint]);
-
-  const isLoading = useMemo(() => {
-    const map = {
-      browser: false,
-      external: isLoadingExternal,
-    };
-    return map[textToSpeechEndpoint];
-  }, [isLoadingExternal, textToSpeechEndpoint]);
+  const isLoading = false;
 
   const voices: Option[] | string[] = useMemo(() => {
-    const voiceMap = {
-      browser: voicesLocal,
-      external: voicesExternal,
-    };
-
-    return voiceMap[textToSpeechEndpoint];
-  }, [textToSpeechEndpoint, voicesExternal, voicesLocal]);
+    return voicesLocal;
+  }, [voicesLocal]);
 
   useEffect(() => {
     const firstVoice = voices[0];
